Внесены исправления в задачу BJS2-66366
