+++ src/main/java/faang/school/analytics/dto/subscription/FollowerEventDto.java
@@ -13,8 +13,8 @@ import java.time.LocalDateTime;
  */
 @Data
 public class FollowerEventDto {
-    private long followerId;
-    private long followeeId;
+    private Long followerId;
+    private Long followeeId;
 
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSSSS")
     private LocalDateTime timestamp;
