+++ src/main/java/faang/school/analytics/service/AnalyticsEventService.java
